Centralize u32 varint reader
